[mtouch] Fail early if the interpreter is used on 32bits
The current interpreter only works on 64 bits (so ARM64). However
the error won't be reported back to the user until deployment time.

This (temporary) fix spot the condition very early and report an error

```
error MT0099 : Internal error : The interpreter is currently only available for 64 bits.
```

instead of the current one at deploy time

```
 IncorrectArchitecture: Failed to find matching arch for 32-bit Mach-O input file /private/var/installd/Library/Caches/com.apple.mobile.installd.staging/temp.tNKDlx/extracted/X.app/X
error MT1006: Could not install the application 'X.app' on the device 'Mercure': AMDeviceSecureInstallApplicationBundle returned: 0xe8000087 (kAMDIncorrectArchitectureError).

Application could not be uploaded to the device.
```

This also centralize other interpreter checks and options in the same
location (making it easier to read / update).